# server/docs/swagger.yaml
definitions:
  entity.NoOrmProjectConfig:
    properties:
      join_by_self:
        description: 是否可以自己加入到顶级任务,默认真（用户肯定不能自己加入到项目中的，必须所有者邀请）
        type: boolean
      must_check_by_other:
        description: 是否必须由其他人检查，默认不需要，
        type: boolean
      secret:
        description: 是否是私密项目，默认不需要
        type: boolean
    type: object
  entity.ProjectStatus:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - ProjectStatusInActive
    - ProjectStatusActive
    - ProjectStatusCompleted
    - ProjectStatusArchived
  entity.RoleType:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    type: integer
    x-enum-varnames:
    - RoleTypeOwner
    - RoleTypeProducter
    - RoleTypeDeveloper
    - RoleTypeTester
    - RoleTypeAdmin
  entity.TaskStatus:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    type: integer
    x-enum-varnames:
    - TaskStatusCreated
    - TaskStarted
    - TaskStatusCompleted
    - TaskStatusArchived
    - TaskFailed
...
  request.ProjectDeleteRole:
    properties:
      project_id:
        type: integer
      user_id:
        type: integer
    type: object
  request.ProjectUpdate:
    properties:
      config:
        $ref: '#/definitions/entity.NoOrmProjectConfig'
      description:
        type: string
      git_address:
        type: string
      id:
        type: integer
      name:
        type: string
    required:
    - id
    type: object
  request.ProjectUpdateStatus:
    properties:
      id:
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/entity.ProjectStatus'
        maximum: 4
        minimum: 0
    required:
    - id
    type: object
  request.ProjectUpsertRole:
    properties:
      project_id:
        type: integer
      role_type:
        $ref: '#/definitions/entity.RoleType'
      user_id:
        type: integer
    type: object
  request.TaskCreateTop:
    properties:
      assign_users:
        items:
          type: integer
        type: array
      description:
        type: string
      expect_complete_time:
        type: string
      name:
        maxLength: 20
        type: string
      project_id:
        minimum: 1
        type: integer
      tester_id:
        minimum: 1
        type: integer
    required:
    - description
    - name
    - project_id
    - tester_id
    type: object
  request.TaskUpdateTop:
    properties:
      assign_users:
        items:
          type: integer
        type: array
      description:
        type: string
      expect_complete_time:
        type: string
      id:
        minimum: 1
        type: integer
      name:
        maxLength: 20
        type: string
      status:
        allOf:
        - $ref: '#/definitions/entity.TaskStatus'
        maximum: 5
        minimum: 1
      tester_id:
        minimum: 1
        type: integer
    required:
    - id
    type: object
  request.UserLogin:
    properties:
      name:
        default: admin
        type: string
      password:
        default: Aa123456
        type: string
      use_mobile:
        type: boolean
    required:
    - name
    - password
    type: object
  request.UserUpdate:
    properties:
      avatar:
        maximum: 20
        minimum: 0
        type: integer
      email:
        type: string
      user_name:
        type: string
    required:
    - avatar
    - email
    - user_name
    type: object
  request.UserUpdatePassword:
    properties:
      new_password:
        minLength: 8
        type: string
      new_password2:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - new_password
    - new_password2
    - old_password
    type: object
  response.CommonResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  response.CommonResponseWithoutData:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  response.Project:
    properties:
      id:
        type: integer
      name:
        type: string
      owner_id:
        type: integer
      owner_name:
        type: string
      role_type:
        $ref: '#/definitions/entity.RoleType'
      status:
        $ref: '#/definitions/entity.ProjectStatus'
    type: object
  response.ProjectList:
    properties:
      list:
        items:
          $ref: '#/definitions/response.Project'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /project/create_role:
    post:
      consumes:
      - application/json
      description: project create role
      parameters:
      - description: AdminRegister
        in: body
        name: ProjectUpsertRole
        required: true
        schema:
          $ref: '#/definitions/request.ProjectUpsertRole'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: project create role
      tags:
      - Project
  /project/delete_role:
    post:
      consumes:
      - application/json
      description: project delete role
      parameters:
      - description: ProjectDeleteRole
        in: body
        name: ProjectDeleteRole
        required: true
        schema:
          $ref: '#/definitions/request.ProjectDeleteRole'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: project delete role
      tags:
      - Project
  /project/info:
    get:
      consumes:
      - '*/*'
      description: project info
      parameters:
      - in: query
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: project info
      tags:
      - Project
  /project/list:
    get:
      consumes:
      - '*/*'
      description: project list
      parameters:
      - default: 1
        in: query
        name: page
        type: integer
      - default: 10
        in: query
        name: page_size
        type: integer
      - default: 0
        in: query
        name: role_type
        type: integer
      - default: 0
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.ProjectList'
      summary: project list
      tags:
      - Project
  /project/role_in:
    get:
      consumes:
      - '*/*'
      description: project role
      parameters:
      - in: query
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: project role in
      tags:
      - Project
  /project/update:
    post:
      consumes:
      - application/json
      description: project update
      parameters:
      - description: ProjectUpdate
        in: body
        name: ProjectUpdate
        required: true
        schema:
          $ref: '#/definitions/request.ProjectUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: project update
      tags:
      - Project
  /project/update_role:
    post:
      consumes:
      - application/json
      description: project update role
      parameters:
      - description: AdminRegister
        in: body
        name: ProjectUpsertRole
        required: true
        schema:
          $ref: '#/definitions/request.ProjectUpsertRole'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: project update role
      tags:
      - Project
  /project/update_status:
    post:
      consumes:
      - application/json
      description: project update status
      parameters:
      - description: ProjectUpdateStatus
        in: body
        name: ProjectUpdateStatus
        required: true
        schema:
          $ref: '#/definitions/request.ProjectUpdateStatus'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: project update status
      tags:
      - Project
  /task/create_top:
    post:
      consumes:
      - application/json
      description: task create top
      parameters:
      - description: AdminRegister
        in: body
        name: TaskCreateTop
        required: true
        schema:
          $ref: '#/definitions/request.TaskCreateTop'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: task create top
      tags:
      - Task
  /task/info:
    get:
      consumes:
      - '*/*'
      description: task info
      parameters:
      - in: query
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: task
      tags:
      - Task
  /task/update_top:
    post:
      consumes:
      - application/json
      description: task update top
      parameters:
      - description: AdminRegister
        in: body
        name: TaskUpdateTop
        required: true
        schema:
          $ref: '#/definitions/request.TaskUpdateTop'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: task update top
      tags:
      - Task
...
  /user/list:
    get:
      consumes:
      - '*/*'
      description: user list
      parameters:
      - in: query
        name: include_admin
        type: boolean
      - in: query
        minimum: 0
        name: project_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: user list
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: user login
      parameters:
      - description: UserRegister
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: user login
      tags:
      - User
...
  /user/update:
    post:
      consumes:
      - application/json
      description: user update1
      parameters:
      - description: UserUpdate
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: user update
      tags:
      - User
  /user/update_password:
    post:
      consumes:
      - application/json
      description: user update_password
      parameters:
      - description: UserUpdatePassword
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.UserUpdatePassword'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.CommonResponseWithoutData'
      summary: user update_password
      tags:
      - User
swagger: "2.0"